# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM ИТ Школы РТК — REST API
  version: 0.2.0
  description: |
    API веб-клиента CRM и точка интеграции с корпоративными системами.

    Аутентификация: вход через Keycloak (GET /api/auth/login) создаёт серверную сессию BFF, браузер получает HttpOnly cookie SESSION; токены OIDC клиенту не передаются. Выход (POST /api/auth/logout) завершает сессию CRM и возвращает адрес завершения сессии Keycloak.

    Изменяющие запросы передают заголовок X-CSRF-TOKEN со значением token из GET /api/csrf. Команды передают Idempotency-Key; повтор с тем же ключом и телом возвращает прежний результат, с другим телом — 409 IDEMPOTENCY_CONFLICT. Устаревшая version даёт 409 VERSION_CONFLICT с currentVersion.

    Ошибки возвращаются в формате ApiError {code, message, requestId}. requestId совпадает с заголовком X-Request-Id и записью в журнале сервера; stack trace и секреты в ответ не попадают.
servers:
  - url: /
...
paths:
...
  /api/organizations:
    get:
      tags: [Catalog]
      summary: Список организаций
      operationId: listOrganizations
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/OrganizationSort'
        - name: q
          in: query
          required: false
          description: Часть названия без учёта регистра
          schema:
            type: string
            maxLength: 200
        - name: requiresAssignment
          in: query
          required: false
          description: Только вузы без ответственного или с ответственным, который не является активным КАМ команды вуза
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Organizations visible within the current server-side business scope, filtered and counted on the server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOrganization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/organizations/{id}:
    get:
      tags: [Catalog]
      summary: Карточка организации
      operationId: getOrganization
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Organization visible within the current server-side business scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/organizations/{id}/contacts:
    get:
      tags: [Catalog]
      summary: Контакты организации
      operationId: listOrganizationContacts
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Contacts belonging to the visible organization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Catalog]
      summary: Добавить контакт организации
      operationId: createOrganizationContact
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Contact created in the visible organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/organizations/{id}/assignment-options:
    get:
      tags: [Catalog]
      summary: Кандидаты в ответственные КАМ
      operationId: listOrganizationAssignmentOptions
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Active USER profiles of the visible organization's team, available only to its LEADER
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationAssignmentCandidate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/organizations/{id}/assignment-events:
    get:
      tags: [Catalog]
      summary: История назначений ответственного
      operationId: listOrganizationAssignmentEvents
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Immutable owner assignment history of the visible organization, available only to its LEADER
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationAssignmentEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/organizations/{id}/assignment:
    post:
      tags: [Catalog]
      summary: Назначить ответственного КАМ
      operationId: assignOrganizationOwner
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationAssignment'
      responses:
        '200':
          description: Updated organization and its immutable owner assignment event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationAssignmentResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: Серверная сессия BFF, созданная входом через GET /api/auth/login. Swagger UI на том же origin использует сессию браузера.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    OrganizationSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: ['name,asc', 'name,desc', 'updatedAt,asc', 'updatedAt,desc']
        default: 'updatedAt,desc'
...
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Значение token из GET /api/csrf для текущей сессии. Без заголовка или с чужим значением запрос отклоняется с 403 FORBIDDEN.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    BadRequest:
      description: Validation or request format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: BFF session is absent or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated identity is not allowed to perform the action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    CrmProfileRequired:
      description: The OIDC identity has no active CRM profile (CRM_PROFILE_REQUIRED) or its profile created at first login waits for administrator activation (CRM_PROFILE_PENDING)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ContactInteractionMutationForbidden:
      description: A visible organization or interaction can be changed only by an active USER or LEADER of its scope; the scope is checked first, so a hidden object is 404. Also returned when the OIDC identity has no active CRM profile
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Object does not exist or is outside the current server-side scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Version or idempotency conflict; VERSION_CONFLICT includes currentVersion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalError:
      description: Unexpected server error; the stack trace is written only to the server log under the same requestId
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHENTICATED
            - FORBIDDEN
            - CRM_PROFILE_REQUIRED
            - CRM_PROFILE_PENDING
            - LAST_ACTIVE_ADMIN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_CONFLICT
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - REPORT_NOT_READY
            - REPORT_ACCESS_CHANGED
            - REPORT_RESULT_UNAVAILABLE
            - REPORT_CAPACITY_EXCEEDED
            - SOURCE_NOT_CONFIGURED
            - SYNC_ALREADY_RUNNING
            - SYNC_CAPACITY_EXCEEDED
            - LMS_NOT_MAPPED
            - LMS_SYNC_FAILED
        message:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        currentVersion:
          type: integer
          minimum: 0
...
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
    Organization:
      type: object
      required: [id, name, type, teamId, version, updatedAt, ownerManagerName, teamName, requiresAssignment]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [UNIVERSITY, SCHOOL]
        teamId:
          type: string
          format: uuid
        ownerManagerId:
          type: [string, 'null']
          format: uuid
        version:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
        ownerManagerName:
          type: [string, 'null']
        teamName:
          type: [string, 'null']
        requiresAssignment:
          type: boolean
          description: true, если ответственного нет или он не является активным КАМ команды вуза
    PageOrganization:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Organization'
    OrganizationAssignmentCandidate:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
    OrganizationAssignment:
      type: object
      required: [version, ownerManagerId]
      properties:
        version:
          type: integer
          minimum: 0
        ownerManagerId:
          description: An active USER of the organization team, or null to explicitly remove the owner
          type: [string, 'null']
          format: uuid
    OrganizationAssignmentEvent:
      type: object
      required: [id, organizationId, commandId, actorProfileId, actorDisplayName, requestId, version, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        commandId:
          type: string
          format: uuid
        previousOwnerManagerId:
          type: [string, 'null']
          format: uuid
        previousOwnerManagerDisplayName:
          type: [string, 'null']
        ownerManagerId:
          type: [string, 'null']
          format: uuid
        newOwnerManagerDisplayName:
          type: [string, 'null']
        actorProfileId:
          type: string
          format: uuid
        actorDisplayName:
          type: string
        requestId:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 0
        occurredAt:
          type: string
          format: date-time
    OrganizationAssignmentResult:
      type: object
      required: [organization, event]
      properties:
        organization:
          $ref: '#/components/schemas/Organization'
        event:
          $ref: '#/components/schemas/OrganizationAssignmentEvent'
    Contact:
      type: object
      required: [id, organizationId, name, version, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        position:
          type: [string, 'null']
        email:
          type: [string, 'null']
          format: email
        phone:
          type: [string, 'null']
        version:
          type: integer
          minimum: 0
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContactCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        position:
          type: [string, 'null']
          maxLength: 200
        email:
          type: [string, 'null']
          format: email
          maxLength: 320
        phone:
          type: [string, 'null']
          maxLength: 50
...
security:
  - cookieAuth: []